Move the patch to lib so we can require it properly
