+++ config/initializers/00_datadog.rb
@@ -1,7 +1,7 @@
 require "simple_server_extensions"
 require "ddtrace"
 require "datadog/statsd"
-require "00_datadog_extensions"
+require "extensions/datadog_sidekiq_patch"
 
 # Allow running via an ENV var (for development usage, for example) ...otherwise
 # exclude some envs by default
