+++ lib/extensions/datadog_sidekiq_patch.rb
